Informatics_Journal/requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
barcode==1.0.4
barcode-generator==0.1rc15
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.51.0
graphviz==0.20.3
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
Levenshtein==0.25.1
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.25.2
opencv-python==4.8.1.78
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.1
playsound==1.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycap==2.6.0
Pygments==2.18.0
pyobject==1.2.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
pyzbar==0.1.9
pyzmq==26.0.3
rapidfuzz==3.9.0
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
semantic-version==2.10.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
treepoem==3.24.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
